Decouple cgo dependencies
(cherry picked from commit 0c5e378d57eb24bab688a1afbbab7eb44d881aa6)
