+++ internal/api/api.go
@@ -1,0 +1,50 @@
+package api
+
+import (
+	"github.com/CosmWasm/wasmvm/types"
+	dbm "github.com/tendermint/tm-db"
+)
+
+type Gas = uint64
+
+// GasMeter is a copy of an interface declaration from cosmos-sdk
+// https://github.com/cosmos/cosmos-sdk/blob/18890a225b46260a9adc587be6fa1cc2aff101cd/store/types/gas.go#L34
+type GasMeter interface {
+	GasConsumed() Gas
+}
+
+/****** DB ********/
+
+// KVStore copies a subset of types from cosmos-sdk
+// We may wish to make this more generic sometime in the future, but not now
+// https://github.com/cosmos/cosmos-sdk/blob/bef3689245bab591d7d169abd6bea52db97a70c7/store/types/store.go#L170
+type KVStore interface {
+	Get(key []byte) []byte
+	Set(key, value []byte)
+	Delete(key []byte)
+
+	// Iterator over a domain of keys in ascending order. End is exclusive.
+	// Start must be less than end, or the Iterator is invalid.
+	// Iterator must be closed by caller.
+	// To iterate over entire domain, use store.Iterator(nil, nil)
+	Iterator(start, end []byte) dbm.Iterator
+
+	// ReverseIterator over a domain of keys in descending order. End is exclusive.
+	// Start must be less than end, or the Iterator is invalid.
+	// Iterator must be closed by caller.
+	ReverseIterator(start, end []byte) dbm.Iterator
+}
+
+type Querier = types.Querier
+
+/***** GoAPI *******/
+
+type (
+	HumanizeAddress     func([]byte) (string, uint64, error)
+	CanonicalizeAddress func(string) ([]byte, uint64, error)
+)
+
+type GoAPI struct {
+	HumanAddress     HumanizeAddress
+	CanonicalAddress CanonicalizeAddress
+}
