+++ internal/api/lib.go
@@ -32,8 +32,6 @@ type Cache struct {
 	ptr *C.cache_t
 }
 
-type Querier = types.Querier
-
 func InitCache(dataDir string, supportedFeatures string, cacheSize uint32, instanceMemoryLimit uint32) (Cache, error) {
 	dataDirBytes := []byte(dataDir)
 	supportedFeaturesBytes := []byte(supportedFeatures)
