+++ lib.go
@@ -1,3 +1,5 @@
+//go:build cgo
+
 package cosmwasm
 
 import (
