+++ lib_nocgo.go
@@ -1,0 +1,26 @@
+//go:build !cgo
+
+package cosmwasm
+
+import (
+	"github.com/CosmWasm/wasmvm/internal/api"
+	"github.com/CosmWasm/wasmvm/types"
+)
+
+// Checksum represents a hash of the Wasm bytecode that serves as an ID. Must be generated from this library.
+type Checksum []byte
+
+// WasmCode is an alias for raw bytes of the wasm compiled code
+type WasmCode []byte
+
+// KVStore is a reference to some sub-kvstore that is valid for one instance of a code
+type KVStore = api.KVStore
+
+// GoAPI is a reference to some "precompiles", go callbacks
+type GoAPI = api.GoAPI
+
+// Querier lets us make read-only queries on other modules
+type Querier = types.Querier
+
+// GasMeter is a read-only version of the sdk gas meter
+type GasMeter = api.GasMeter
